Keep this PR to community quotes UI; move blogs and FAQ schema out.
The sourced blog blockquotes and FAQPage JSON-LD will land in a follow-up.
