+++ app/(v1)/compare-to-temporal/page.tsx
@@ -2,11 +2,6 @@ import { type Metadata } from "next";
 import { generateMetadata } from "src/utils/social";
 
 import CompareToTemporal from "@/components/v1/pages/CompareToTemporal";
-import { COMPARE_TEMPORAL_FAQS } from "@/components/v1/sections/CompareTemporal/Faq";
-import {
-  JsonLdScript,
-  buildFaqPageSchema,
-} from "@/utils/v1/structuredData";
 
 export const metadata: Metadata = generateMetadata({
   title: "Inngest vs Temporal: Durable execution that developers love",
@@ -15,17 +10,5 @@ export const metadata: Metadata = generateMetadata({
 });
 
 export default function Page() {
-  const faqSchema = buildFaqPageSchema(
-    COMPARE_TEMPORAL_FAQS.map((faq) => ({
-      question: faq.question,
-      answer: faq.answer,
-    })),
-  );
-
-  return (
-    <>
-      <JsonLdScript data={faqSchema} />
-      <CompareToTemporal />
-    </>
-  );
+  return <CompareToTemporal />;
 }
