+++ content/blog/build-more-reliable-workflows-with-events.mdx
@@ -5,11 +5,9 @@ subtitle: Run critical code with guarantees and observability
 image: /assets/blog/build-more-reliable-workflows-with-events.jpg
 imageCredits: Photo by <a href="https://unsplash.com/@charlesdeluvio?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText">charlesdeluvio</a> on <a href="https://unsplash.com/?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText">Unsplash</a>
 date: 2022-11-10
-dateUpdated: 2026-08-18
+dateUpdated: 2026-01-07
 author: Dan Farrelly
 ---
-
-import Blockquote from "src/shared/Blog/Blockquote";
 
 Every application starts off with running some basic code in the background. Common use cases include sending email notifications, handling subscription payment webhooks, or periodically fetching data from third party APIs.
 
@@ -46,12 +44,6 @@ Our goals were to enable developers to write complex jobs combining multiple ste
 We don't think writing step functions or workflows as a large JSON configuration is a great developer experience. Also, using a online GUI to design these isn't always ideal as the logic or [DAG](https://en.wikipedia.org/wiki/Directed_acyclic_graph) may be edited or managed outside of the actual code's version control.
 
 We also believe it should be simple to run, from local development, to production where code can be [deployed to any number of platforms](/docs/apps/cloud?ref=blog-build-more-reliable-workflows-with-events).
-
-<Blockquote
-  text="Integrated Inngest for background jobs. First event: feature.request.created. This single event triggers the entire AI pipeline — decouples the API from AI execution completely. Retry support built in. No lost jobs."
-  attribution={{ name: "Tejas A.", title: "Developer" }}
-  sourceUrl="https://x.com/TEJAS_DEV_code/status/2070888066378838324"
-/>
 
 Let's take a use case to demonstrate our approach. A product is a CRM tool that enables a user to upload a large CSV file with contacts information. The API triggers this to run with the event `api/contact_list.uploaded`. This advanced CRM tool validates the upload then uses third party APIs to enrich the data with information about the contact's business then waits for the user to review and filter this list down before inserting all data into the CRM database. When the user approve or rejects and selects filers, the API will send another event: `api/contact_list.reviewed`. Here's what the code could look like:
