+++ shared/Blog/Blockquote.tsx
@@ -9,9 +9,6 @@ type BlockquoteProps = {
   };
   avatar?: string;
   logo?: string;
-  /** Link to the original post (X, LinkedIn, Reddit). */
-  sourceUrl?: string;
-  sourceLabel?: string;
 };
 
 export default function Blockquote({
@@ -19,38 +16,22 @@ export default function Blockquote({
   attribution,
   avatar,
   logo,
-  sourceUrl,
-  sourceLabel = "View original",
 }: BlockquoteProps) {
   return (
-    <figure className="not-prose rounded-lg border border-indigo-300/20 py-6 pl-8 pr-10">
+    <figure className="not-prose pl-8 pr-10 py-6 rounded-lg border border-indigo-300/20">
       <blockquote className="text-lg">&ldquo;{text}&rdquo;</blockquote>
-      <div className="mt-6 flex flex-col gap-4 sm:flex-row sm:items-center">
+      <div className="mt-6 flex flex-col sm:flex-row sm:items-center gap-4">
         {!!avatar && (
           <img
-            className="h-10 w-10 rounded-full"
+            className="rounded-full h-10 w-10"
             src={avatar}
             alt={`Image of ${attribution.name}`}
           />
         )}
         <figcaption>
-          <span className="font-semibold text-white">{attribution.name}</span> -{" "}
+          <span className="text-white font-semibold">{attribution.name}</span> -{" "}
           {attribution.title}
           {attribution.company && `, ${attribution.company}`}
-          {sourceUrl ? (
-            <>
-              {" "}
-              ·{" "}
-              <a
-                href={sourceUrl}
-                className="text-indigo-300 hover:text-indigo-200"
-                rel="noopener noreferrer"
-                target="_blank"
-              >
-                {sourceLabel}
-              </a>
-            </>
-          ) : null}
         </figcaption>
         {!!logo && (
           <div className="flex grow sm:justify-end">
